+++ apps/patient/src/api/index.ts
@@ -1,0 +1,2 @@
+export * from './external';
+export * from './internal';
