+++ apps/patient/src/graphql/queries.ts
@@ -1,9 +1,5 @@
 import { gql } from 'graphql-request';
 
-/**
- * Queries
- */
-
 export const GET_ORDER = gql`
   query order($id: ID!) {
     order(id: $id) {
@@ -89,33 +85,5 @@ export const GET_PHARMACIES = gql`
       }
       distance
     }
-  }
-`;
-
-/**
- * Mutations
- */
-
-export const MARK_ORDER_AS_PICKED_UP = gql`
-  mutation MarkOrderAsPickedUp($markOrderAsPickedUpId: ID!) {
-    markOrderAsPickedUp(id: $markOrderAsPickedUpId)
-  }
-`;
-
-export const SELECT_ORDER_PHARMACY = gql`
-  mutation SelectOrderPharmacy($orderId: ID!, $pharmacyId: String, $patientId: String) {
-    selectOrderPharmacy(orderId: $orderId, pharmacyId: $pharmacyId, patientId: $patientId)
-  }
-`;
-
-export const SET_PREFERRED_PHARMACY = gql`
-  mutation SetPreferredPharmacy($pharmacyId: String, $patientId: String) {
-    setPreferredPharmacy(pharmacyId: $pharmacyId, patientId: $patientId)
-  }
-`;
-
-export const REROUTE_ORDER = gql`
-  mutation RerouteOrder($orderId: ID!, $pharmacyId: String) {
-    rerouteOrder(orderId: $orderId, pharmacyId: $pharmacyId)
   }
 `;
